chore(maven): convert project to maven
