+++ .gitignore
@@ -1,3 +1,4 @@
 /bin/
 .DS_Store
 /src/config/assistant.properties
+/target/
