--- src/main/constants/Task.java
@@ -1,10 +1,0 @@
-package main.constants;
-
-/**
- * Enum of values describing task that can be performed by the assistant
- * @author Gbenga Oshinaga
- *
- */
-public enum Task {
-	SEARCH, DICTIONARY, CONVERSATION;
-}
