--- src/main/dictionarymodel/Entry.java
@@ -1,34 +1,0 @@
-
-package main.dictionarymodel;
-
-import java.util.ArrayList;
-import java.util.List;
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Entry {
-
-    @SerializedName("homographNumber")
-    @Expose
-    private String homographNumber;
-    @SerializedName("senses")
-    @Expose
-    private List<Sense> senses = new ArrayList<Sense>();
-
-    public String getHomographNumber() {
-        return homographNumber;
-    }
-
-    public void setHomographNumber(String homographNumber) {
-        this.homographNumber = homographNumber;
-    }
-
-    public List<Sense> getSenses() {
-        return senses;
-    }
-
-    public void setSenses(List<Sense> senses) {
-        this.senses = senses;
-    }
-
-}
