--- src/main/dictionarymodel/Metadata.java
@@ -1,21 +1,0 @@
-
-package main.dictionarymodel;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Metadata {
-
-    @SerializedName("provider")
-    @Expose
-    private String provider;
-
-    public String getProvider() {
-        return provider;
-    }
-
-    public void setProvider(String provider) {
-        this.provider = provider;
-    }
-
-}
