+++ src/main/java/config/assistant.properties
@@ -1,0 +1,4 @@
+DICTIONARY_APP_ID=99b10a8b
+DICTIONARY_KEY=8e6253268d7f45e15bc158966eb0d596
+SEARCH_API_KEY=AIzaSyCU0KqylrmVtXxTXkrq7snY-G5kvg-cqsI
+SEARCH_ENGINE_ID=013668024517642414081:frv-brj1sfy
