+++ src/main/java/constants/Task.java
@@ -1,0 +1,10 @@
+package constants;
+
+/**
+ * Enum of values describing task that can be performed by the assistant
+ * @author Gbenga Oshinaga
+ *
+ */
+public enum Task {
+	SEARCH, DICTIONARY, CONVERSATION;
+}
