+++ src/main/java/dictionarymodel/DictionaryModel.java
@@ -1,0 +1,34 @@
+
+package dictionarymodel;
+
+import java.util.ArrayList;
+import java.util.List;
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class DictionaryModel {
+
+    @SerializedName("metadata")
+    @Expose
+    private Metadata metadata;
+    @SerializedName("results")
+    @Expose
+    private List<Result> results = new ArrayList<Result>();
+
+    public Metadata getMetadata() {
+        return metadata;
+    }
+
+    public void setMetadata(Metadata metadata) {
+        this.metadata = metadata;
+    }
+
+    public List<Result> getResults() {
+        return results;
+    }
+
+    public void setResults(List<Result> results) {
+        this.results = results;
+    }
+
+}
