+++ src/main/java/dictionarymodel/LexicalEntry.java
@@ -1,0 +1,56 @@
+
+package dictionarymodel;
+
+import java.util.ArrayList;
+import java.util.List;
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class LexicalEntry {
+
+    @SerializedName("entries")
+    @Expose
+    private List<Entry> entries = new ArrayList<Entry>();
+    @SerializedName("language")
+    @Expose
+    private String language;
+    @SerializedName("lexicalCategory")
+    @Expose
+    private String lexicalCategory;
+    @SerializedName("text")
+    @Expose
+    private String text;
+
+    public List<Entry> getEntries() {
+        return entries;
+    }
+
+    public void setEntries(List<Entry> entries) {
+        this.entries = entries;
+    }
+
+    public String getLanguage() {
+        return language;
+    }
+
+    public void setLanguage(String language) {
+        this.language = language;
+    }
+
+    public String getLexicalCategory() {
+        return lexicalCategory;
+    }
+
+    public void setLexicalCategory(String lexicalCategory) {
+        this.lexicalCategory = lexicalCategory;
+    }
+
+    public String getText() {
+        return text;
+    }
+
+    public void setText(String text) {
+        this.text = text;
+    }
+
+}
