+++ src/main/java/dictionarymodel/Metadata.java
@@ -1,0 +1,21 @@
+
+package dictionarymodel;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class Metadata {
+
+    @SerializedName("provider")
+    @Expose
+    private String provider;
+
+    public String getProvider() {
+        return provider;
+    }
+
+    public void setProvider(String provider) {
+        this.provider = provider;
+    }
+
+}
