+++ src/test/java/mock/JsonMock.java
@@ -1,0 +1,95 @@
+package mock;
+
+public interface JsonMock {
+  public final String goatJson = "{\n" + 
+      "    \"metadata\": {\n" + 
+      "        \"provider\": \"Oxford University Press\"\n" + 
+      "    },\n" + 
+      "    \"results\": [\n" + 
+      "        {\n" + 
+      "            \"id\": \"goat\",\n" + 
+      "            \"language\": \"en\",\n" + 
+      "            \"lexicalEntries\": [\n" + 
+      "                {\n" + 
+      "                    \"entries\": [\n" + 
+      "                        {\n" + 
+      "                            \"homographNumber\": \"000\",\n" + 
+      "                            \"senses\": [\n" + 
+      "                                {\n" + 
+      "                                    \"definitions\": [\n" + 
+      "                                        \"a hardy domesticated ruminant mammal that has backward-curving horns and (in the male) a beard. It is kept for its milk and meat, and noted for its lively behaviour.\"\n" + 
+      "                                    ],\n" + 
+      "                                    \"id\": \"m_en_gbus0419580.007\",\n" + 
+      "                                    \"subsenses\": [\n" + 
+      "                                        {\n" + 
+      "                                            \"definitions\": [\n" + 
+      "                                                \"a wild mammal related to the goat, such as the ibex.\"\n" + 
+      "                                            ],\n" + 
+      "                                            \"id\": \"m_en_gbus0419580.011\"\n" + 
+      "                                        },\n" + 
+      "                                        {\n" + 
+      "                                            \"definitions\": [\n" + 
+      "                                                \"the zodiacal sign Capricorn or the constellation Capricornus.\"\n" + 
+      "                                            ],\n" + 
+      "                                            \"id\": \"m_en_gbus0419580.012\"\n" + 
+      "                                        }\n" + 
+      "                                    ]\n" + 
+      "                                },\n" + 
+      "                                {\n" + 
+      "                                    \"definitions\": [\n" + 
+      "                                        \"a lecherous man.\"\n" + 
+      "                                    ],\n" + 
+      "                                    \"id\": \"m_en_gbus0419580.016\"\n" + 
+      "                                },\n" + 
+      "                                {\n" + 
+      "                                    \"definitions\": [\n" + 
+      "                                        \"a stupid person; a fool\"\n" + 
+      "                                    ],\n" + 
+      "                                    \"id\": \"m_en_gbus0419580.018\"\n" + 
+      "                                },\n" + 
+      "                                {\n" + 
+      "                                    \"definitions\": [\n" + 
+      "                                        \"a scapegoat.\"\n" + 
+      "                                    ],\n" + 
+      "                                    \"id\": \"m_en_gbus0419580.020\"\n" + 
+      "                                }\n" + 
+      "                            ]\n" + 
+      "                        }\n" + 
+      "                    ],\n" + 
+      "                    \"language\": \"en\",\n" + 
+      "                    \"lexicalCategory\": \"Noun\",\n" + 
+      "                    \"text\": \"goat\"\n" + 
+      "                },\n" + 
+      "                {\n" + 
+      "                    \"entries\": [\n" + 
+      "                        {\n" + 
+      "                            \"homographNumber\": \"001\",\n" + 
+      "                            \"senses\": [\n" + 
+      "                                {\n" + 
+      "                                    \"definitions\": [\n" + 
+      "                                        \"(especially of a sports player) greatest of all time\"\n" + 
+      "                                    ],\n" + 
+      "                                    \"id\": \"m_en_gbus0419590.003\"\n" + 
+      "                                }\n" + 
+      "                            ]\n" + 
+      "                        }\n" + 
+      "                    ],\n" + 
+      "                    \"language\": \"en\",\n" + 
+      "                    \"lexicalCategory\": \"Residual\",\n" + 
+      "                    \"text\": \"GOAT\"\n" + 
+      "                }\n" + 
+      "            ],\n" + 
+      "            \"type\": \"headword\",\n" + 
+      "            \"word\": \"goat\"\n" + 
+      "        }\n" + 
+      "    ]\n" + 
+      "}";
+  
+  public final String goatDefinitons = "~ a hardy domesticated ruminant mammal that has backward-curving horns and (in the male) a beard. It is kept for its milk and meat, and noted for its lively behaviour.\n" + 
+      "~ a wild mammal related to the goat, such as the ibex.\n" + 
+      "~ the zodiacal sign Capricorn or the constellation Capricornus.\n" + 
+      "~ a lecherous man.\n" + 
+      "~ a stupid person; a fool\n" + 
+      "~ a scapegoat.\n" + 
+      "";
+}
